+++ lib/core/src/system/gpu/data/texture/types/format.rs
@@ -1,9 +1,8 @@
-/// Texture formats. A `GlEnum` specifying the format of the texel data. Follow the link to learn
-/// more: https://developer.mozilla.org/en-US/docs/Web/API/WebGLRenderingContext/texImage2D
+//! Texture formats. A `GlEnum` specifying the format of the texel data. Follow the link to learn
+//! more: https://developer.mozilla.org/en-US/docs/Web/API/WebGLRenderingContext/texImage2D
 
 use crate::prelude::*;
 
-use crate::system::gpu::data::prim::*;
 use crate::system::gpu::data::texture::gl_enums::*;
 use crate::system::gpu::data::gl_enum::GlEnum;
