+++ lib/core/src/system/gpu/data/texture/types/internal_format.rs
@@ -1,5 +1,5 @@
-/// A GLenum specifying the color components in the texture. Follow the link to learn more:
-/// https://developer.mozilla.org/en-US/docs/Web/API/WebGLRenderingContext/texImage2D
+//! A GLenum specifying the color components in the texture. Follow the link to learn more:
+//! https://developer.mozilla.org/en-US/docs/Web/API/WebGLRenderingContext/texImage2D
 
 use crate::prelude::*;
 
@@ -8,11 +8,8 @@ use crate::system::gpu::data::gl_enum::*;
 use crate::system::gpu::data::texture::gl_enums::*;
 use crate::system::gpu::data::texture::format::*;
 use crate::system::gpu::data::texture::sampler::*;
-use crate::system::gpu::data::texture::Value;
 
 use nalgebra::*;
-
-use crate::system::gpu::data::texture::{True,False};
